chore: api is now versioned and front end has been extracted to its own module
