--- api/api_gateway/templates/dashboard.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <!--Import Google Icon Font-->
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-  ​
-  <!-- Compiled and minified CSS -->
-  <link href="https://unpkg.com/tailwindcss@^2/dist/tailwind.min.css" rel="stylesheet" />
-  <title>Scan Websites</title>
-</head>
-​
-<body>
-  {% block content %}
-  Organizations
-  <tr>
-  {% for organization in organisations %}
-    <td>{{organization.name}}</td>
-  {% endfor %}
-  </tr>
-  {% endblock %}
-  
-</body>
-
-</html>
