+++ api/front_end/templates/layout.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Scan websites</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://unpkg.com/tailwindcss@^2/dist/tailwind.min.css" rel="stylesheet" />
+    <style type="text/css">
+      .container {
+        max-width: 500px;
+        padding-top: 100px;
+      }
+      h2 {color: red;}
+    </style>
+  </head>
+  <body>
+    <div class="container">
+      <h2>TODO: HEADER</h2>
+      <br>
+      {% block content %}{% endblock %}
+      <br>
+      <div class="footer">
+        {% block footer %}
+          TODO: Footer goes here
+          <br>
+          <br>
+        {% endblock %}
+      </div>
+    </div>
+  </body>
+</html>
