+++ api/front_end/templates/page.html
@@ -1,0 +1,7 @@
+{% extends "layout.html" %}
+{% block content %}
+  <div>Welcome to Scan websites</div>
+  {% block footer %}
+  {{super()}}
+  {% endblock %}
+{% endblock %}
