+++ bin/api_healthcheck.sh
@@ -3,7 +3,7 @@
 echo hitting api healthcheck endpoint
 curl "http://api:8080/2015-03-31/functions/function/invocations" -d '{
     "resource": "/",
-    "path": "/healthcheck",
+    "path": "/api/v1/healthcheck",
     "requestContext": {},
     "httpMethod": "GET",
     "headers": {},
