+++ bin/api_version.sh
@@ -3,7 +3,7 @@
 echo hitting api version endpoint
 curl "http://api:8080/2015-03-31/functions/function/invocations" -d '{
     "resource": "/",
-    "path": "/version",
+    "path": "/api/v1/version",
     "requestContext": {},
     "httpMethod": "GET",
     "headers": {},
